FasterThanJson: changed length and size variables to be int instead of uint. Added Decimal to fast tuple writer and reader.

+++ src/Starcounter.XSON/Sessions/ManagedTupleWriter.cs
@@ -68,7 +68,7 @@ namespace Starcounter.Internal {
          buffer = null;
          unsafe {
             fixed (byte* pbuf = buffer) {
-               tw = new TupleWriterBase64(pbuf, (uint)valueCount, (uint)initialOffsetElementSize);
+               tw = new TupleWriterBase64(pbuf, (uint)valueCount, initialOffsetElementSize);
             }
          }
       }
